requirements.txt
python-dotenv==0.19.0
pyTelegramBotAPI==4.14.0